- Added a dependency for the utility project - Added a rough table structure to handle the new calendar interface - Added Dao's based on the old implementation to save these objects to the database

+++ ProductionEJB/src/main/java/com/danielbchapman/production/entity/EntityInstance.java
@@ -33,7 +33,8 @@ public class EntityInstance
 	 */
 	public static void deleteObject(org.theactingcompany.persistence.Indentifiable obj)
 	{
-		getDelegate().deleteObject(obj);
+	  if(obj != null)
+	    getDelegate().deleteObject(obj);
 	}
 
 	/**
